+++ R/extract_coef.R
@@ -1,5 +1,5 @@
 # Standardized coefficient extraction
-# 
+#
 # Extract model coefficients in a standardized format (as a named numeric vector).
 #
 # @section Warning:
@@ -8,9 +8,9 @@
 # internally recoded to -1 and +1. sqlscore multiplies the returned coefficients by 2
 # to put them back on the same scale as glm, and adds the glmboost offset to the
 # intercept before multiplying.
-# 
+#
 # @param object An object for which the extraction of model coefficients is meaningful.
-# 
+#
 # @return Model coefficients as a named numeric vector.
 extract_coef <-
 function(object)
@@ -29,29 +29,25 @@ function(object)
 extract_coef.glmboost <-
 function(object)
 {
-  families <- c("Negative Binomial Likelihood -- probit Link",
-                "Negative Binomial Likelihood", "Poisson Likelihood", 
-                "Squared Error (Regression)", "Negative Gamma Likelihood")
-  stopifnot(object$family@name %in% families)
-  
   # suppress coef.glmboost's message so we can print our own
   cf <- captureConditions(stats::coef(object, off2int=TRUE))$value
-  
-  # mboost internally recodes binomial DVs to -1 and +1, so the coefficients
-  # are half those returned by glm. If we have a binomial model, let's fix this
-  # and return twice the fitted coefficients, with a message
-  if(object$family@name %in% families[1:2])
+
+  # mboost internally recodes DVs to -1 and +1 in one particular case (the
+  # logit fit from Binomial_adaboost()), so the coefficients are half those
+  # returned by glm. If we have a binomial model, let's fix this and return
+  # twice the fitted coefficients, with a message
+  if(object$family@name == "Negative Binomial Likelihood (logit link)")
   {
-    message("\nNOTE: Coefficients from a glmboost Binomial model are 1/2 the ",
-            "coefficients\nfrom a model fit by ", "glm(... , family = 'binomial').\n", 
+    message("\nNOTE: Coefficients from glmboost's Binomial_adaboost logit model \n",
+            "are 1/2 the coefficients from a model fit by glm(... , family = 'binomial').\n",
             "sqlscore scales these coefficients by 2 to put them on the same scale as glm.\n")
     sc <- 2
   }
   else
   {
-    sc <- 1 
+    sc <- 1
   }
-  
+
   sc * cf
 }
 
@@ -60,10 +56,10 @@ extract_coef.cv.glmnet <-
 function(object)
 {
   cf <- stats::coef(object)
-  
+
   val <- as.vector(cf)
   names(val) <- rownames(cf)
-  
+
   #Return only the coefficients that weren't regularized to 0
   val[which(val != 0)]
 }