+++ tests/testthat/test-fqtn.R
@@ -1,18 +1,20 @@
 context("Table name escaping")
 
+sql <- get_sql()
+
 test_that("Basic table names are escaped", {
-  expect_equal(fqtn("tbl"), dplyr::sql('"tbl"'))
-  expect_equal(fqtn("foo", schema="bar"), dplyr::sql('"bar"."foo"'))
-  expect_equal(fqtn("foo", schema="bar", catalog="baz"), dplyr::sql('"baz"."bar"."foo"'))
+  expect_equal(fqtn("tbl"), sql('"tbl"'))
+  expect_equal(fqtn("foo", schema="bar"), sql('"bar"."foo"'))
+  expect_equal(fqtn("foo", schema="bar", catalog="baz"), sql('"baz"."bar"."foo"'))
 })
 
 test_that("More complicated names are escaped", {
-  expect_equal(fqtn("foo bar"), dplyr::sql('"foo bar"'))
-  expect_equal(fqtn("foo bar", schema="baz"), dplyr::sql('"baz"."foo bar"'))
+  expect_equal(fqtn("foo bar"), sql('"foo bar"'))
+  expect_equal(fqtn("foo bar", schema="baz"), sql('"baz"."foo bar"'))
   expect_equal(fqtn("foo bar", schema="baz", catalog="quux quux"),
-               dplyr::sql('"quux quux"."baz"."foo bar"'))
+               sql('"quux quux"."baz"."foo bar"'))
   expect_equal(fqtn("foo", schema="baz", catalog="quux quux"),
-               dplyr::sql('"quux quux"."baz"."foo"'))
+               sql('"quux quux"."baz"."foo"'))
 })
 
 test_that("Invalid calls throw errors", {
@@ -23,4 +25,3 @@ test_that("Invalid calls throw errors", {
   expect_error(fqtn("foo", catalog="bar")) #need schema
   expect_error(fqtn(schema="foo", catalog="bar")) #need table
 })
-